# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Full Express API",
    "version": "1.0.0",
    "description": "A comprehensive Express API with Prisma"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://back-end-production-17b6.up.railway.app/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/certificates": {
      "get": {
        "summary": "Get all certificates",
        "tags": [
          "Certificates"
        ],
        "responses": {
          "200": {
            "description": "List of all certificates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Certificate"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new certificate",
        "tags": [
          "Certificates"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Certificate created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/certificates/{id}": {
      "get": {
        "summary": "Get a certificate by ID",
        "tags": [
          "Certificates"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Certificate ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "404": {
            "description": "Certificate not found"
          }
        }
      },
      "put": {
        "summary": "Update a certificate",
        "tags": [
          "Certificates"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Certificate ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Certificate updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            }
          },
          "404": {
            "description": "Certificate not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a certificate",
        "tags": [
          "Certificates"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Certificate ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Certificate deleted successfully"
          },
          "404": {
            "description": "Certificate not found"
          }
        }
      }
    },
...
    "/external-api-keys": {
      "post": {
        "summary": "Crear una nueva API Key externa",
        "tags": [
          "ExternalApiKey"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nombre del sistema externo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API Key creada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "active": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar todas las API Keys externas",
        "tags": [
          "ExternalApiKey"
        ],
        "responses": {
          "200": {
            "description": "Lista de API Keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "key": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "revokedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/external-api-keys/{id}/revoke": {
      "patch": {
        "summary": "Revocar una API Key externa",
        "tags": [
          "ExternalApiKey"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la API Key"
          }
        ],
        "responses": {
          "200": {
            "description": "API Key revocada"
          },
          "404": {
            "description": "API Key no encontrada"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Certificate": {
        "type": "object",
        "required": [
          "studentId",
          "courseId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the certificate"
          },
          "studentId": {
            "type": "string",
            "description": "ID of the student who earned the certificate"
          },
          "courseId": {
            "type": "string",
            "description": "ID of the course the certificate is for"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Certificate issuance timestamp"
          },
          "certificateUrl": {
            "type": "string",
            "description": "URL to the certificate document"
          }
        }
      },
      "CertificateInput": {
        "type": "object",
        "required": [
          "studentId",
          "courseId"
        ],
        "properties": {
          "studentId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "certificateUrl": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}